[JENKINS-36687] migrate "cleanSlashes" into generic utils
