--- blueocean-core-js/src/js/sse/UrlUtils.js
@@ -1,20 +1,0 @@
-/**
- * Created by cmeyers on 8/15/16.
- */
-
-/**
- * Trims duplicate forward slashes to a single slash and adds trailing slash if needed.
- * @param url
- * @returns {string}
- */
-export const cleanSlashes = (url) => {
-    if (url.indexOf('//') !== -1) {
-        let cleanUrl = url.replace('//', '/');
-        cleanUrl = cleanUrl.substr(-1) === '/' ?
-            cleanUrl : `${cleanUrl}/`;
-
-        return cleanSlashes(cleanUrl);
-    }
-
-    return url;
-};
