+++ blueocean-core-js/src/js/utils.js
@@ -1,6 +1,24 @@
 
 
+/**
+ * Trims duplicate forward slashes to a single slash and adds trailing slash if needed.
+ * @param url
+ * @returns {string}
+ */
+const cleanSlashes = (url) => {
+    if (url.indexOf('//') !== -1) {
+        let cleanUrl = url.replace('//', '/');
+        cleanUrl = cleanUrl.substr(-1) === '/' ?
+            cleanUrl : `${cleanUrl}/`;
+
+        return cleanSlashes(cleanUrl);
+    }
+
+    return url;
+};
+
 export default {
+    cleanSlashes,
     clone(obj) {
         return JSON.parse(JSON.stringify(obj));
     },
